move GraphUtils code to its own translation unit
