+++ src/ast/analysis/PrecedenceGraph.h
@@ -21,7 +21,9 @@
 #include "GraphUtils.h"
 #include "ast/Relation.h"
 #include "ast/TranslationUnit.h"
-#include <string>
+#include "souffle/datastructure/Graph.h"
+#include <ostream>
+#include <sstream>
 
 namespace souffle::ast {
