+++ src/ast/analysis/typesystem/TypeEnvironment.cpp
@@ -15,7 +15,6 @@
  ***********************************************************************/
 
 #include "ast/analysis/typesystem/TypeEnvironment.h"
-#include "GraphUtils.h"
 #include "ast/AlgebraicDataType.h"
 #include "ast/AliasType.h"
 #include "ast/Attribute.h"
@@ -27,6 +26,7 @@
 #include "ast/Type.h"
 #include "ast/UnionType.h"
 #include "ast/analysis/typesystem/TypeSystem.h"
+#include "souffle/datastructure/Graph.h"
 #include "souffle/utility/MiscUtil.h"
 #include "souffle/utility/tinyformat.h"
 #include <algorithm>
