+++ src/ast/transform/InlineRelations.cpp
@@ -42,7 +42,9 @@
 #include "ast/UnnamedVariable.h"
 #include "ast/UserDefinedFunctor.h"
 #include "ast/Variable.h"
+#include "ast/analysis/Ground.h"
 #include "ast/analysis/typesystem/PolymorphicObjects.h"
+#include "ast/analysis/typesystem/TypeEnvironment.h"
 #include "ast/utility/Utils.h"
 #include "ast/utility/Visitor.h"
 #include "souffle/BinaryConstraintOps.h"
